rename: routes 내 test를 TestRoutes로 이름 변경 및 적용(#46)
