+++ src/App.tsx
@@ -1,6 +1,6 @@
-import TestApp from './routes/test'
+import TestRoutes from './routes/TestRoutes'
 // import MainApp from "./routes/main"; // 메인이 필요해지면 교체
 
 export default function App() {
-  return <TestApp />
+  return <TestRoutes />
 }
